src/create_thumbnail_image/requirements.txt
functions-framework==3.*
google-cloud-retail
firebase-functions==0.1.*
pillow
requests